make.sh: Mount /usr/src in memory for faster build
This patch enables that /usr/src is a ramdisk which should
give us fewer I/O operations when extracting tarballs or
writing small intermediate files by the compiler.

In some virtualised environments this should bring a good
performance boost.

There is no persistent data stored in this directory and
some persistent directories are mounted over it.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>