udev: Set wireless interfaces into AP mode before adding to bridge
Wireless interfaces cannot be added to the bridge when they are
still running in station mode. At boot time, the interface will
be created and it is attempted to attach it to a bridge, which
fails.

This patch now checks whether the wireless interface is going to
be an access point, changes the mode accordingly and then attaches
the interface to the bridge.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>
Tested-by: Daniel Weismüller <daniel.weismueller@ipfire.org>
Signed-off-by: Arne Fitzenreiter <arne_f@ipfire.org>