Package the official Go compiler
Since Go has a horrible build system which requires a Go
compiler to build the Go compiler and takes a very long
time to compile, we are following Rust and are using the
"official" pre-compiled release tarball.

We no longer ship the Go runtime, which mitigates the
risk of shipping any malware.

Because we currently only have one package using this
and which is only being compiled for x86_64, we are
only making Go available on this architecture.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>
Signed-off-by: Arne Fitzenreiter <arne_f@ipfire.org>