Update boost to 1.71.0
I did not update to the current version 1.74.0 because another package
I want to submit do not build with the current version.

I do not think this is a big show stopper because other distributions
are not on the latest version either.

The removed switch do not seem to be availavle in this version anymore.

Signed-off-by: Jonatan Schlag <jonatan.schlag@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>