kernel: enable ACPI support on ARM64
ACPI (with EFI) is used on ARM systems conforming to the
Server Base Boot Requirements (SBBR) and is an optional
on embedded systems (EBBR).

Up to now the ARM64 boards supported by IPFire use U-Boot and
device tree so ACPI was not turned on.

The immediate use case here is to run under virtualization,
using my muvirt project[1] I can run IPFire on our Traverse Ten64
system. For reasons I'll explain separately it is not
currently possible to run stock IPFire on this system.

This change also enables the EFI RTC driver which is presented
by the qemu arm64 virt machine.

Signed-off-by: Mathew McBride <matt@traverse.com.au>

[1] - https://gitlab.com/traversetech/muvirt
Signed-off-by: Arne Fitzenreiter <arne_f@ipfire.org>