ipinfo.cgi: Display network flags of the given addresses.
Use the libloc data for gathering and displaying the stored network
flags, like "Anonymous Proxy" for the addresses.

The notice of a flag only will be displayed, if a flag is set for the
network which contains the given address.

Currently this notice text is "hardcoded" in englisch language, because
the entire other content of the page is in Englisch (responses from RIR's)
and also the flag names like "Anonymous Proxy" are only availabe in
English.

IMHO there is no need to to translate the string "This address is marked
as" into different languages, because of the reasons abouve.

Signed-off-by: Stefan Schantl <stefan.schantl@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>