kernel: Trust the randomness from the CPU
This will allow the kernel to seed its CRNG using RDSEED or RDRAND.

During the boot process, it is required that the CRNG is being
initialised, but it may take some long time on systems that do not have
a random number generator.

This is the default for various other distributions like Debian.

Signed-off-by: Arne Fitzenreiter <arne.fitzenreiter@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>