make.sh: Do not decrease nice level of build
This was designed to keep a workstation that is compiling IPFire
responsive during the build. However, the kernel's scheduler has been
improved enough that this is no longer an issue.

Instead of telling the kernel that the build job is something with a
lower priority (which it isn't) we now simply run with the nicelevel of
the parent process that has called make.sh.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>